+++ changelog.md
@@ -8,6 +8,7 @@
 -   [d]—docs improvement
 
 ## 2.7.1 (March XX, 2023)
+- [+] Bulk SQLite and PostgreSQL update with single query!y
 - [+] Enum Support
 
 ## 2.7.0 (February 27, 2023)